Panel Administrator.
Made - User, Manufacturer, Category, Supplier control.

Method view all list, create new, delete, edit.

+++ src/main/java/com/vint/shop/repository/OrderProductMapRepository.java
@@ -3,5 +3,5 @@ package com.vint.shop.repository;
 import com.vint.shop.domain.OrderProductMap;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderProductMapRepository extends JpaRepository<OrderProductMap, Integer> {
+public interface OrderProductMapRepository extends JpaRepository<OrderProductMap, Long> {
 }
